Use our own OpenSSL build
Apple has deprecated linking to the system provided OpenSSL libraries. In this commit, I added a target that downloads and builds our own copy of OpenSSL. I've also updated the postgresql build script to compile and link against our version of OpenSSL. Additionally, I updated the dylib-target so that it also fixes all the OpenSSL-related paths.