use "postgres" as username during initdb #275
This change uses the user name "postgres" when calling initdb. This is the de facto standard used by other PostgreSQL distributions, and should make restoring from dumps more usable.

To keep Postgres.app as simple to use, Postgres.app also creates a super user with the same name as the current system user, and creates a database with the same name. This makes sure that typing "psql" in terminal is all that's needed to connect to Postgres.app