Always register login item on macOS 13 and newer
This fixes an issue where macOS 13 showed a warning that "Jakob Egger" installed software to run in the background (issue #677).
The toggle in the preferences is removed as well.
If the user does not want Postgres.app to start on login, they can now disable it in System Settings.
I think this way is best, because otherwise there would be multiple places to change this setting.