Add PostgresPermissionDialog
PostgresPermissionDialog is an executable that can be called
by postmaster to show a macOS style permission dialog.

It communicates the users reply via exit code:
- Exit Code 0: allow
- Exit Code 1: deny
- Other exit codes: error

Error Messages are printed to stderr.
My plan is to have postmaster call the executable with system().
I hope that error messages will then appear in the PostgreSQL
server log.