devel: allow to build from git
Instead of specifying a version for the tarball download, we can pass
a git URL with a hash or ref from which the tarball is built. This
minimises changes to the postgresql make target while allowing builds
from HEAD, a cf_bot branch or a custom fork. The builds are put in a
'devel' prefix to keep them separate from current and future release
versions. When build from git the hash is included in extra-version.