Fix version string parsing and clean up persisted garbage
In 5e444c5 we added a "(Postgres.app)" suffix to the version string,
causing Server.binaryVersion to return this instead of the actual
version. As a result, garbage was written to postgresapp_config.plist
and servers initialized with pg15.6-17.2 did not use ICU as their
default collation provider and were missing data-checksums.