pg13-14: update OpenSSL to 3.0 LTS
To ensure a smooth transition, enable the legacy provider by default.
For this, we create a OPENSSLDIR within our bundle, place a proper
openssl.cnf and replace the solution introduced in 6f20a05by by a
symlink to the system provided certs. OpenSSL Engines and Providers
will now be included in distribution.